// src/payload-types.ts
export interface Media {
  id: string;
  /**
   * Alt text voor toegankelijkheid
   */
  alt?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    thumbnail?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    card?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    tablet?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "projects".
 */
export interface Project {
  id: string;
  /**
   * Project naam
   */
  Naam: string;
  /**
   * Is dit project actief en zichtbaar?
   */
  Actief?: boolean | null;
  /**
   * Aantal verkochte units
   */
  Aantalverkocht?: number | null;
  /**
   * Type project (bijv. Appartementen, Huizen)
   */
  Type?: string | null;
  /**
   * Aantal units van dit type
   */
  Aantalvantype?: number | null;
  /**
   * Locatie van het project
   */
  Locatie?: string | null;
  /**
   * Huidige fase van het project
   */
  Fase?:
    | (
        | "planning"
        | "ruwbouw in uitvoering"
        | "verkocht als projectgrond"
        | "in afwerking"
        | "opgeleverd"
        | "afgewerkt"
      )
    | null;
  /**
   * Prijs informatie (bijv. "Vanaf €250.000")
   */
  Prijs?: string | null;
  /**
   * Gedetailleerde projectomschrijving
   */
  Projectomschrijving?: string | null;
  /**
   * Geografische locatie van het project (longitude, latitude)
   *
   * @minItems 2
   * @maxItems 2
   */
  LocatieCoordinaten?: [number, number] | null;
  /**
   * Tabel data (JSON string)
   */
  Tabel?: string | null;
  /**
   * Hoofdfoto van het project
   */
  Hoofdfoto: string | Media;
  /**
   * Afbeelding met prijsinformatie
   */
  prijsImage?: (string | null) | Media;
  /**
   * Fotogalerij van het project
   */
  Fotosvhp?:
    | {
        foto: string | Media;
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
}